+++ src/main/java/insane96mcp/enhancedai/modules/animal/AnimalScaredAttack.java
@@ -8,11 +8,12 @@ import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
-import insane96mcp.insanelib.base.config.Blacklist;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.data.IdTagMatcher;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
@@ -20,43 +21,29 @@ import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
 import net.minecraft.world.entity.ai.goal.PanicGoal;
-import net.minecraft.world.entity.ai.goal.TemptGoal;
 import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
-import net.minecraft.world.entity.ai.targeting.TargetingConditions;
 import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.monster.Enemy;
 import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.event.entity.EntityAttributeModificationEvent;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
-import net.minecraftforge.event.entity.living.LivingDamageEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-import java.util.List;
-
-@Label(name = "Animals", description = "Make animals fight back or group flee when attacked")
+@Label(name = "Animals Scared Attack", description = "Make animals fight back or be scared by players. Use the entity type tag enhancedai:can_fight_back and enhancedai:can_be_scared_by_players to change animals")
 @LoadFeature(module = Modules.Ids.ANIMAL)
-public class Animals extends Feature {
-
+public class AnimalScaredAttack extends Feature {
+    public static final TagKey<EntityType<?>> CAN_FIGHT_BACK = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "can_fight_back"));
+    public static final TagKey<EntityType<?>> CAN_BE_SCARED_BY_PLAYERS = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "can_be_scared_by_players"));
     public static final String CAN_ATTACK_BACK = EnhancedAI.RESOURCE_PREFIX + "can_attack_back";
     public static final String PLAYER_SCARED = EnhancedAI.RESOURCE_PREFIX + "player_scared";
-    public static final String NOT_TEMPTED = EnhancedAI.RESOURCE_PREFIX + "not_tempted";
 
-    @Config
-    @Label(name = "Group Flee", description = "If true, when an animal is attacked, all the animals around will flee.")
-    public static Boolean groupFlee = true;
-    @Config
-    @Label(name = "Flee Range", description = "If Group Flee is enabled, this is the range where the animals will flee.")
-    public static Integer groupFleeRange = 16;
     @Config
     @Label(name = "Fight back chance", description = "Animals have this percentage chance to be able to fight back instead of fleeing. Animals have a slightly bigger range to attack. Attack damage can't be changed via config due to limitation so use mods like Mobs Properties Randomness to change the damage. Base damage is 3.")
     public static Double fightBackChance = 0.3d;
     @Config
     @Label(name = "Players Scared chance", description = "Animals have this percentage chance to be scared by players and run away. Fight back chance has priority over this.")
     public static Double playersScaredChance = 0.4d;
-    @Config
-    @Label(name = "Not tempted chance", description = "Animals have this percentage chance to not be temped by food.")
-    public static Double notTemptedChance = 0.5d;
     @Config(min = 0d, max = 4d)
     @Label(name = "Movement Speed Multiplier", description = "Movement speed multiplier when aggroed.")
     public static Double speedMultiplier = 1.3d;
@@ -64,20 +51,12 @@ public class Animals extends Feature {
     @Label(name = "Knockback", description = "Animals' knockback attribute will be set to this value.")
     public static Double knockback = 1.5d;
     @Config
-    @Label(name = "Entity Blacklist", description = "Entities that shouldn't be affected by this feature")
-    public static Blacklist entityBlacklist = new Blacklist(List.of(
-            IdTagMatcher.newId("minecraft:llama"),
-            IdTagMatcher.newId("minecraft:trader_llama"),
-            IdTagMatcher.newId("minecraft:bee"),
-            IdTagMatcher.newId("minecraft:wolf"),
-            IdTagMatcher.newId("minecraft:bear"),
-            IdTagMatcher.newId("minecraft:panda"),
-            IdTagMatcher.newId("minecraft:axolotl")
-    ), false);
+    @Label(name = "Knockback size based", description = "Animals' knockback attribute will be increased/decreased based on the side of the mob.")
+    public static Boolean knockbackSizeBased = true;
 
     private static final double BASE_ATTACK_DAMAGE = 3d;
 
-    public Animals(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+    public AnimalScaredAttack(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
     }
 
@@ -95,47 +74,32 @@ public class Animals extends Feature {
     public void onSpawn(EntityJoinLevelEvent event) {
         if (!this.isEnabled()
                 || event.getEntity() instanceof Enemy
-                || !(event.getEntity() instanceof Animal animal)
-                || entityBlacklist.isEntityBlackOrNotWhitelist(animal))
+                || !(event.getEntity() instanceof Animal animal))
             return;
 
         CompoundTag persistentData = animal.getPersistentData();
 
         double movementSpeedMultiplier = NBTUtils.getDoubleOrPutDefault(persistentData, EATags.Passive.SPEED_MULTIPLIER_WHEN_AGGROED, speedMultiplier);
-        boolean canAttackBack = NBTUtils.getBooleanOrPutDefault(persistentData, CAN_ATTACK_BACK, animal.getRandom().nextDouble() < fightBackChance);
-        boolean playerScared = NBTUtils.getBooleanOrPutDefault(persistentData, PLAYER_SCARED, animal.getRandom().nextDouble() < playersScaredChance);
-        boolean notTempted = NBTUtils.getBooleanOrPutDefault(persistentData, NOT_TEMPTED, animal.getRandom().nextDouble() < notTemptedChance);
+        boolean canAttackBack = NBTUtils.getBooleanOrPutDefault(persistentData, CAN_ATTACK_BACK, animal.getType().is(CAN_FIGHT_BACK) && animal.getRandom().nextDouble() < fightBackChance);
+        boolean playerScared = NBTUtils.getBooleanOrPutDefault(persistentData, PLAYER_SCARED, !canAttackBack && animal.getType().is(CAN_BE_SCARED_BY_PLAYERS) && animal.getRandom().nextDouble() < playersScaredChance);
 
         if (canAttackBack && !animal.isBaby()) {
             animal.targetSelector.addGoal(1, (new HurtByTargetGoal(animal)).setAlertOthers());
             animal.goalSelector.addGoal(1, new MeleeAttackGoal(animal, movementSpeedMultiplier, true));
             animal.goalSelector.availableGoals.removeIf(wrappedGoal -> wrappedGoal.getGoal() instanceof PanicGoal);
             if (knockback > 0d) {
+                double baseSize = 1.053d; // Sheep square meters size
+                double actualKnockback = knockback;
+                if (knockbackSizeBased)
+                    actualKnockback = (animal.getBbWidth() * animal.getBbWidth() * animal.getBbHeight()) * knockback / baseSize;
                 AttributeInstance kbAttribute = animal.getAttribute(Attributes.ATTACK_KNOCKBACK);
                 if (kbAttribute != null)
-                    kbAttribute.addPermanentModifier(new AttributeModifier("Animal knockback", knockback, AttributeModifier.Operation.ADDITION));
+                    kbAttribute.addPermanentModifier(new AttributeModifier("Animal knockback", actualKnockback, AttributeModifier.Operation.ADDITION));
             }
         }
         else if (playerScared) {
             EAAvoidEntityGoal<Player> avoidEntityGoal = new EAAvoidEntityGoal<>(animal, Player.class, (float) 16, (float) 8, 1.25, 1.1);
             animal.goalSelector.addGoal(1, avoidEntityGoal);
         }
-
-        if (notTempted) {
-            animal.goalSelector.getAvailableGoals().removeIf(wrappedGoal -> wrappedGoal.getGoal() instanceof TemptGoal);
-        }
-    }
-
-    @SubscribeEvent
-    public void onAttacked(LivingDamageEvent event) {
-        if (!this.isEnabled()
-                || !groupFlee
-                || !(event.getEntity() instanceof Animal animal)
-                || entityBlacklist.isEntityBlackOrNotWhitelist(animal)
-                || !(event.getSource().getEntity() instanceof LivingEntity attacker))
-            return;
-
-        animal.level().getNearbyEntities(Animal.class, TargetingConditions.forNonCombat().ignoreLineOfSight(), animal, animal.getBoundingBox().inflate(groupFleeRange))
-                .stream().filter(otherAnimal -> otherAnimal.getType().equals(animal.getType())).forEach(nearbyAnimal -> nearbyAnimal.setLastHurtByMob(attacker));
     }
 }